Complete main navigate
