+++ restaurant/src/main/java/bus/AuthenticationBus.java
@@ -3,11 +3,6 @@ package bus;
 import dao.Repository;
 import dao.exceptions.AuthenticationFailException;
 import dao.exceptions.FetchUserFailException;
-import javafx.scene.Scene;
-import ui.base.StageManager;
-import ui.signInScreen.SignInScreen;
-
-import java.io.IOException;
 
 public class AuthenticationBus {
     private Repository _repository = Repository.getInstance();
@@ -23,10 +18,5 @@ public class AuthenticationBus {
     public void logout() {
         System.out.println("Logout...");
         _repository.logout();
-        try {
-            StageManager.getInstance().pushReplacement(new Scene(new SignInScreen()), "Đăng nhập");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 }
