+++ restaurant/src/main/java/ui/base/StageManager.java
@@ -1,10 +1,18 @@
 package ui.base;
 
-import javafx.event.EventHandler;
+import com.jfoenix.controls.JFXDecorator;
+import de.jensd.fx.glyphs.fontawesome.FontAwesomeIcon;
+import de.jensd.fx.glyphs.fontawesome.FontAwesomeIconView;
+import io.datafx.controller.flow.Flow;
+import io.datafx.controller.flow.FlowException;
+import io.datafx.controller.flow.FlowHandler;
+import io.datafx.controller.flow.container.ContainerAnimations;
+import javafx.collections.ObservableList;
 import javafx.scene.Scene;
+import javafx.scene.layout.StackPane;
+import javafx.scene.paint.Color;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
-import javafx.stage.WindowEvent;
+import javafx.util.Duration;
 
 import java.util.Stack;
 
@@ -12,20 +20,12 @@ public class StageManager {
     //region Singleton
     private static StageManager _ourInstance = new StageManager();
     //endregion
-    private Stack<Stage> _mainStages;
-    private boolean _isShowingStage;
-    private EventHandler<WindowEvent> closeStageEvent = new EventHandler<WindowEvent>() {
-        @Override
-        public void handle(WindowEvent event) {
-            closeAndPopCurrent();
-            if (!_isShowingStage)
-                showCurrent();
-        }
-    };
+    private Stack<Stage> _stageStack;
+    private Stack<FlowHandler> _flowHandlerStack;
 
     private StageManager() {
-        _mainStages = new Stack<>();
-        _isShowingStage = false;
+        _stageStack = new Stack<>();
+        _flowHandlerStack = new Stack<>();
     }
 
     public static StageManager getInstance() {
@@ -33,64 +33,106 @@ public class StageManager {
     }
 
     public Stage getCurrent() {
-        return _mainStages.peek();
+        return _stageStack.peek();
     }
 
-    public void push(Stage stage) {
-        _isShowingStage = true;
-        closeCurrent();
-        addNewStage(stage);
+    public void push(Class<?> startViewControllerClass, double width, double height, String title, Stage primaryStage) throws FlowException {
+        // Create new stage.
+        Stage newStage = prepareNewStage(primaryStage, startViewControllerClass, width, height, title);
+
+        // Close current stage.
+        if (_stageStack.size() > 0) _stageStack.peek().close();
+
+        // Push stage to stack
+        _stageStack.push(newStage);
+
+        // Show top stage of stack
+        showTopStack();
     }
 
-    public void pushReplacement(Stage stage) {
-        _isShowingStage = true;
-        closeAndPopCurrent();
-        addNewStage(stage);
-    }
+    public void pushReplacement(Class<?> startViewControllerClass, double width, double height, String title, Stage primaryStage) throws FlowException {
+        // Create new stage.
+        Stage newStage = prepareNewStage(primaryStage, startViewControllerClass, width, height, title);
 
-    public void push(Scene scene, String title) {
-        Stage newStage = new Stage();
-        newStage.initStyle(StageStyle.UTILITY);
-        newStage.setTitle(title);
-        newStage.setResizable(false);
-        newStage.setScene(scene);
-        push(newStage);
-    }
+        // Close current stage and pop stack.
+        if (_stageStack.size() > 0) {
+            _stageStack.peek().close();
+            popStack();
+        }
 
-    public void pushReplacement(Scene scene, String title) {
-        Stage newStage = new Stage();
-        newStage.initStyle(StageStyle.UTILITY);
-        newStage.setTitle(title);
-        newStage.setResizable(false);
-        newStage.setScene(scene);
-        pushReplacement(newStage);
+        // Push stage to stack
+        _stageStack.push(newStage);
+
+        // Show top stage of stack
+        showTopStack();
     }
 
     public void pop() {
-        closeAndPopCurrent();
-    }
-
-    private void addNewStage(Stage stage) {
-        _mainStages.push(stage);
-        stage.setOnCloseRequest(closeStageEvent);
-        showCurrent();
-    }
-
-    private void closeAndPopCurrent() {
-        if (_mainStages.size() > 0) {
-            _mainStages.pop().close();
+        // Close current stage and pop stack.
+        if (_stageStack.size() > 0) {
+            _stageStack.peek().close();
+            popStack();
         }
+
+        // Show top stage of stack
+        showTopStack();
     }
 
-    private void closeCurrent() {
-        if (_mainStages.size() > 0) {
-            _mainStages.peek().close();
-        }
+    public FlowHandler getCurrentFlowHandler(){
+        return _flowHandlerStack.peek();
     }
 
-    private void showCurrent() {
-        if (_mainStages.size() > 0)
-            _mainStages.peek().show();
-        _isShowingStage = false;
+    private void onStageClosed() {
+        popStack();
+        showTopStack();
+    }
+
+    private void popStack() {
+        _stageStack.pop();
+        _flowHandlerStack.pop();
+    }
+
+    /**
+     * Create new stage, decorate and add on closed listener to it.
+     *
+     * @param startViewControllerClass Controller class of screen.
+     * @param width                    Window width.
+     * @param height                   Window height.
+     * @param title                    Window title.
+     * @return New window.
+     */
+    private Stage prepareNewStage(Stage stage, Class<?> startViewControllerClass, double width, double height, String title) throws FlowException {
+        if (stage == null)
+            stage = new Stage();
+
+        // Create flow
+        Flow flow = new Flow(startViewControllerClass);
+        ExtendedAnimatedFlowContainer container = new ExtendedAnimatedFlowContainer(Duration.millis(500), ContainerAnimations.ZOOM_OUT);
+        FlowHandler flowHandler = flow.createHandler();
+        _flowHandlerStack.push(flowHandler);
+        flowHandler.start(container);
+
+
+        // Decorate stage
+        JFXDecorator decorator = new JFXDecorator(stage, container.getView());
+        decorator.setCustomMaximize(true);
+        FontAwesomeIconView icon = new FontAwesomeIconView(FontAwesomeIcon.CUTLERY);
+        icon.setFill(Color.WHITE);
+        decorator.setGraphic(icon);
+
+        stage.setTitle(title);
+        Scene scene = new Scene(new StackPane(decorator), width, height);
+        final ObservableList<String> stylesheets = scene.getStylesheets();
+        stylesheets.addAll(StageManager.class.getResource("../../css/jfoenix-main-demo.css").toExternalForm());
+        stage.setScene(scene);
+
+        //Add listener
+        stage.setOnCloseRequest(event -> onStageClosed());
+
+        return stage;
+    }
+
+    private void showTopStack() {
+        if (_stageStack.size() > 0) _stageStack.peek().show();
     }
 }
