--- restaurant/src/main/java/ui/mainScreen/Controller.java
@@ -1,72 +1,0 @@
-package ui.mainScreen;
-
-import com.jfoenix.controls.JFXDrawer;
-import com.jfoenix.controls.JFXHamburger;
-import com.jfoenix.controls.JFXRippler;
-import com.jfoenix.controls.JFXTooltip;
-import io.datafx.controller.ViewController;
-import io.datafx.controller.flow.Flow;
-import io.datafx.controller.flow.FlowHandler;
-import io.datafx.controller.flow.context.FXMLViewFlowContext;
-import io.datafx.controller.flow.context.ViewFlowContext;
-import javafx.animation.Transition;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.layout.StackPane;
-import javafx.util.Duration;
-import ui.accountScreen.AccountScreen;
-
-import java.net.URL;
-import java.util.ResourceBundle;
-
-@ViewController(value = "/ui/mainScreen/test.fxml")
-public class Controller implements Initializable {
-    @FXMLViewFlowContext
-    private ViewFlowContext context;
-    @FXML
-    private JFXDrawer drawer;
-    @FXML
-    private JFXHamburger titleBurger;
-    @FXML
-    private StackPane titleBurgerContainer;
-    @FXML
-    public JFXRippler optionsRippler;
-    @FXML
-    public StackPane optionsBurger;
-
-    @Override
-    public void initialize(URL location, ResourceBundle resources) {
-        final JFXTooltip burgerTooltip = new JFXTooltip("Open drawer");
-
-        drawer.setOnDrawerOpening(e -> {
-            final Transition animation = titleBurger.getAnimation();
-            burgerTooltip.setText("Close drawer");
-            animation.setRate(1);
-            animation.play();
-        });
-        drawer.setOnDrawerClosing(e -> {
-            final Transition animation = titleBurger.getAnimation();
-            burgerTooltip.setText("Open drawer");
-            animation.setRate(-1);
-            animation.play();
-        });
-        titleBurgerContainer.setOnMouseClicked(e -> {
-            if (drawer.isClosed() || drawer.isClosing()) {
-                drawer.open();
-            } else {
-                drawer.close();
-            }
-        });
-
-        // create the inner flow and content
-        context = new ViewFlowContext();
-        // set the default controller
-        Flow innerFlow = new Flow(AccountScreen.class);
-
-        final FlowHandler flowHandler = innerFlow.createHandler(context);
-        context.register("ContentFlowHandler", flowHandler);
-        context.register("ContentFlow", innerFlow);
-        final Duration containerAnimationDuration = Duration.millis(320);
-
-    }
-}
