+++ restaurant/src/main/resources/ui/mainScreen/MainPopup.fxml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXListView?>
+<?import com.jfoenix.controls.JFXPopup?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.StackPane?>
+<?import org.kordamp.ikonli.javafx.FontIcon?>
+<JFXPopup fx:id="toolbarPopup" styleClass="option-list-view" xmlns="http://javafx.com/javafx/8"
+          xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.mainScreen.MainPopup">
+    <JFXListView fx:id="toolbarPopupList">
+        <Label>
+            Contact Us
+            <graphic>
+                <StackPane styleClass="custom-jfx-list-view-icon-container">
+                    <FontIcon iconLiteral="far-user" iconSize="18" styleClass="option-jfx-list-view-icon"/>
+                </StackPane>
+            </graphic>
+        </Label>
+        <Label onMouseClicked="#logout">
+            Đăng xuất
+            <graphic>
+                <StackPane styleClass="custom-jfx-list-view-icon-container">
+                    <FontIcon iconLiteral="fas-reply" iconSize="18" styleClass="option-jfx-list-view-icon"/>
+                </StackPane>
+            </graphic>
+        </Label>
+    </JFXListView>
+</JFXPopup>
