+++ restaurant/src/main/resources/ui/splashScreen/SplashScreen.fxml
@@ -1,15 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import com.jfoenix.controls.JFXProgressBar?>
+<?import javafx.geometry.Insets?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane prefHeight="346.0" prefWidth="600.0" style="-fx-background-color: linear-gradient(from 0% 0% to 100% 100%, #FC5C7D, #6A82FB);" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.splashScreen.Controller">
-    <ImageView fx:id="logo" fitHeight="150.0" fitWidth="200.0" layoutX="225.0" layoutY="98.0" pickOnBounds="true" preserveRatio="true">
-        <image>
-          <Image url="@../../images/logo.png" />
-        </image>
+<?import javafx.scene.layout.FlowPane?>
+<FlowPane alignment="CENTER" columnHalignment="CENTER" orientation="VERTICAL" prefHeight="346.0" prefWidth="600.0"
+          style="-fx-background-color: linear-gradient(from 0% 50% to 100% 50%, #FC5C7D, #9c27b0);"
+          xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1">
+    <ImageView fx:id="logo" fitHeight="150.0" fitWidth="200.0" layoutX="225.0" layoutY="98.0" pickOnBounds="true"
+               preserveRatio="true">
+        <Image url="@../../images/logo.png"/>
     </ImageView>
-   <JFXProgressBar fx:id="progressBar" visible="false" layoutX="200.0" layoutY="275.0" prefHeight="11.0" />
-</AnchorPane>
+    <JFXProgressBar fx:id="progressBar" layoutX="200.0" layoutY="275.0" prefHeight="11.0" visible="false">
+        <FlowPane.margin>
+            <Insets top="50.0"/>
+        </FlowPane.margin>
+    </JFXProgressBar>
+</FlowPane>
